Revise HTTP Client stub
