+++ .github/workflows/ci.yml
@@ -66,10 +66,10 @@ jobs:
       MINIO_ACCESS_KEY: sourcemeta-test
       MINIO_SECRET_KEY: 12345678
       MINIO_BUCKET: test
+      HTTP_STUB_PORT: 9999
 
       CC: ${{ matrix.platform.cc }}
       CXX: ${{ matrix.platform.cxx }}
-      TEST_PORT: 9999
     steps:
       - name: Install ClangFormat
         run: pip install clang-format==18.1.0
@@ -140,10 +140,10 @@ jobs:
         run: (minio server --address :${{ env.MINIO_PORT }} build/minio &) && sleep 10 && make configure-minio
         if: runner.os != 'windows'
       - name: Run HTTP client stub (Windows)
-        run: cmd /c "start /b node test\stub.js ${{ env.TEST_PORT }}" && Start-Sleep -Seconds 10
+        run: cmd /c "start /b node test\e2e\httpclient\stub.js ${{ env.HTTP_STUB_PORT }}" && Start-Sleep -Seconds 10
         if: runner.os == 'windows'
       - name: Run HTTP client stub (*nix)
-        run: (node test/stub.js ${{ env.TEST_PORT }} &) && sleep 10
+        run: (node test/e2e/httpclient/stub.js ${{ env.HTTP_STUB_PORT }} &) && sleep 10
         if: runner.os != 'windows'
 
       # Not every CTest version supports the --test-dir option. If such option
@@ -154,8 +154,7 @@ jobs:
         run: cd ./build && ctest --build-config Release --output-on-failure --parallel
         if: runner.os != 'windows'
         env:
-          # TODO: Rename this variable and get rid of TEST_PORT
-          SOURCEMETA_HYDRA_TEST_SERVER_BASE_URL: 'http://localhost:${{ env.TEST_PORT }}'
+          SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL: 'http://localhost:${{ env.HTTP_STUB_PORT }}'
           SOURCEMETA_HYDRA_TEST_BUCKET_BASE_URL: ${{ env.MINIO_PROTOCOL }}://${{ env.MINIO_ADDRESS }}:${{ env.MINIO_PORT }}/${{ env.MINIO_BUCKET }}
           SOURCEMETA_HYDRA_TEST_BUCKET_REGION: ${{ env.MINIO_REGION }}
           SOURCEMETA_HYDRA_TEST_BUCKET_ACCESS_KEY: ${{ env.MINIO_ACCESS_KEY }}
@@ -169,5 +168,4 @@ jobs:
         run: cd ./build && ctest --build-config Release --exclude-regex e2e_Bucket --output-on-failure --parallel
         if: runner.os == 'windows'
         env:
-          # TODO: Rename this variable and get rid of TEST_PORT
-          SOURCEMETA_HYDRA_TEST_SERVER_BASE_URL: 'http://localhost:${{ env.TEST_PORT }}'
+          SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL: 'http://localhost:${{ env.HTTP_STUB_PORT }}'
