+++ Makefile
@@ -10,7 +10,6 @@ MKDIR = mkdir
 # Options
 PRESET = Debug
 SHARED = OFF
-TEST_PORT = 9999
 
 all: configure compile test
 
@@ -84,13 +83,14 @@ configure-minio: .always
 		--secret-key $(MINIO_SECRET_KEY) \
 		$(MINIO_ALIAS) $(MINIO_ROOT_USER)
 
-test-stub: test/stub.js .always
-	$(NODE) $< $(TEST_PORT)
+HTTP_STUB_PORT = 9999
+up-http-stub: test/e2e/httpclient/stub.js .always
+	$(NODE) $< $(HTTP_STUB_PORT)
 
-test-e2e: configure compile configure-minio 
+test-e2e: configure compile configure-minio
 	$(CMAKE) -E env \
 		UBSAN_OPTIONS=print_stacktrace=1 \
-		SOURCEMETA_HYDRA_TEST_SERVER_BASE_URL=http://localhost:$(TEST_PORT) \
+		SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL=http://localhost:$(HTTP_STUB_PORT) \
 		SOURCEMETA_HYDRA_TEST_BUCKET_BASE_URL=$(MINIO_PROTOCOL)://$(MINIO_ADDRESS):$(MINIO_PORT)/$(MINIO_BUCKET) \
 		SOURCEMETA_HYDRA_TEST_BUCKET_REGION=$(MINIO_REGION) \
 		SOURCEMETA_HYDRA_TEST_BUCKET_ACCESS_KEY=$(MINIO_ACCESS_KEY) \
