+++ test/e2e/httpclient/environment.h
@@ -1,0 +1,24 @@
+#ifndef SOURCEMETA_HYDRA_TEST_HTTPCLIENT_ENVIRONMENT_H_
+#define SOURCEMETA_HYDRA_TEST_HTTPCLIENT_ENVIRONMENT_H_
+
+#include <cstdlib>  // std::exit, std::getenv
+#include <iostream> // std::cerr
+#include <string>   // std::string
+
+static const char *const SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL =
+    std::getenv("SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL");
+
+inline auto HTTPCLIENT_BASE_URL() -> const std::string & {
+  if (!SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL) {
+    std::cerr << "Missing environment variable: "
+                 "SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL\n";
+    std::exit(1);
+  }
+
+  static std::string BASE_URL{SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL};
+  std::cerr << "SOURCEMETA_HYDRA_TEST_HTTPCLIENT_BASE_URL: " << BASE_URL
+            << "\n";
+  return BASE_URL;
+}
+
+#endif
