+++ test/e2e/httpclient/request_1_1_test.cc
@@ -11,7 +11,7 @@
 
 #include <sourcemeta/hydra/httpclient.h>
 
-#include "http_base_url.h"
+#include "environment.h"
 
 static auto
 body(sourcemeta::hydra::http::ClientResponse &response) -> std::string {
@@ -45,7 +45,7 @@ TEST(e2e_HTTP_Request_1_1, method_get_by_default) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -54,7 +54,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root) {
 }
 
 TEST(e2e_HTTP_Request_1_1, HEAD_root) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::HEAD);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -62,7 +62,7 @@ TEST(e2e_HTTP_Request_1_1, HEAD_root) {
 }
 
 TEST(e2e_HTTP_Request_1_1, POST_root) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::POST);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -71,7 +71,7 @@ TEST(e2e_HTTP_Request_1_1, POST_root) {
 }
 
 TEST(e2e_HTTP_Request_1_1, PUT_root) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::PUT);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -80,7 +80,7 @@ TEST(e2e_HTTP_Request_1_1, PUT_root) {
 }
 
 TEST(e2e_HTTP_Request_1_1, DELETE_root) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::DELETE);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -89,7 +89,7 @@ TEST(e2e_HTTP_Request_1_1, DELETE_root) {
 }
 
 TEST(e2e_HTTP_Request_1_1, OPTIONS_root) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::OPTIONS);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -98,7 +98,7 @@ TEST(e2e_HTTP_Request_1_1, OPTIONS_root) {
 }
 
 TEST(e2e_HTTP_Request_1_1, TRACE_root) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::TRACE);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -107,7 +107,7 @@ TEST(e2e_HTTP_Request_1_1, TRACE_root) {
 }
 
 TEST(e2e_HTTP_Request_1_1, PATCH_root) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::PATCH);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -116,8 +116,8 @@ TEST(e2e_HTTP_Request_1_1, PATCH_root) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_foo_bar) {
-  sourcemeta::hydra::http::ClientRequest request{std::string{HTTP_BASE_URL()} +
-                                                 "/foo/bar"};
+  sourcemeta::hydra::http::ClientRequest request{
+      std::string{HTTPCLIENT_BASE_URL()} + "/foo/bar"};
   request.method(sourcemeta::hydra::http::Method::GET);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -126,7 +126,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_foo_bar) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_custom_code_string) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.header("X-Code", "400");
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -136,7 +136,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_custom_code_string) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_custom_code_integer) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.header("X-Code", 400);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -146,7 +146,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_custom_code_integer) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_content_type_no_capture) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -156,7 +156,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_content_type_no_capture) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_content_type_capture) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture("content-type");
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -168,7 +168,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_content_type_capture) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_content_type_capture_all) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture();
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -179,7 +179,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_content_type_capture_all) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_missing_header_with_capture) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture("x-foo-bar");
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -190,7 +190,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_missing_header_with_capture) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_multiple_captures_match) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture("content-type");
   request.capture("content-length");
@@ -204,7 +204,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_multiple_captures_match) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_multiple_captures_partial_match) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture("content-type");
   request.capture("x-foo-bar");
@@ -218,7 +218,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_multiple_captures_partial_match) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_multiple_captures_match_initializer_list) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture({"content-type", "content-length"});
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -232,7 +232,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_multiple_captures_match_initializer_list) {
 
 TEST(e2e_HTTP_Request_1_1,
      GET_root_multiple_captures_partial_match_initializer_list) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture({"content-type", "x-foo-bar"});
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -245,7 +245,7 @@ TEST(e2e_HTTP_Request_1_1,
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_request_header_echo) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.header("x-foo-bar", "foo");
   request.capture("x-x-foo-bar");
@@ -258,7 +258,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_request_header_echo) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_request_header_mixed_case_echo) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.header("X-Foo-Bar", "foo");
   request.capture("x-x-foo-bar");
@@ -271,7 +271,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_request_header_mixed_case_echo) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_multiple_times) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
 
   sourcemeta::hydra::http::ClientResponse response_1{request.send().get()};
@@ -286,7 +286,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_multiple_times) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_move) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   sourcemeta::hydra::http::ClientRequest new_request{std::move(request)};
   sourcemeta::hydra::http::ClientResponse response{new_request.send().get()};
@@ -296,7 +296,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_move) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_headers_capture_none) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -309,7 +309,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_headers_capture_none) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_headers_capture_one) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture("content-type");
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -324,7 +324,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_headers_capture_one) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_headers_capture_all) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture();
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -339,7 +339,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_headers_capture_all) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_header_last_modified_gmt) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture("last-modified");
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -368,7 +368,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_header_last_modified_gmt) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_header_gmt_invalid) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.capture("content-type");
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
@@ -377,7 +377,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_header_gmt_invalid) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_response_header_gmt_missing) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   sourcemeta::hydra::http::ClientResponse response{request.send().get()};
   EXPECT_EQ(response.status(), sourcemeta::hydra::http::Status::OK);
@@ -386,7 +386,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_response_header_gmt_missing) {
 }
 
 TEST(e2e_HTTP_Request_1_1, POST_root_hello_world) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::POST);
   std::istringstream request_body{"hello world"};
   sourcemeta::hydra::http::ClientResponse response{
@@ -397,7 +397,7 @@ TEST(e2e_HTTP_Request_1_1, POST_root_hello_world) {
 }
 
 TEST(e2e_HTTP_Request_1_1, GET_root_hello_world) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   std::istringstream request_body{"hello world"};
   sourcemeta::hydra::http::ClientResponse response{
@@ -408,7 +408,7 @@ TEST(e2e_HTTP_Request_1_1, GET_root_hello_world) {
 }
 
 TEST(e2e_HTTP_Request_1_1, PUT_root_hello_world) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::PUT);
   std::istringstream request_body{"hello world"};
   sourcemeta::hydra::http::ClientResponse response{
@@ -419,7 +419,7 @@ TEST(e2e_HTTP_Request_1_1, PUT_root_hello_world) {
 }
 
 TEST(e2e_HTTP_Request_1_1, root_empty_istringstream) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::POST);
   std::istringstream request_body;
   sourcemeta::hydra::http::ClientResponse response{
@@ -430,7 +430,7 @@ TEST(e2e_HTTP_Request_1_1, root_empty_istringstream) {
 }
 
 TEST(e2e_HTTP_Request_1_1, unsigned_long_header) {
-  sourcemeta::hydra::http::ClientRequest request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientRequest request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   const unsigned long number{11};
   request.header("x-foo", number);
