+++ test/e2e/httpclient/stream_1_1_test.cc
@@ -8,7 +8,7 @@
 
 #include <sourcemeta/hydra/httpclient.h>
 
-#include "http_base_url.h"
+#include "environment.h"
 
 TEST(e2e_HTTP_Stream_1_1, invalid_url) {
   sourcemeta::hydra::http::ClientStream request{"foobarbaz"};
@@ -32,22 +32,22 @@ TEST(e2e_HTTP_Stream_1_1, method_get_by_default) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, no_callbacks_get) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   const auto status{request.send().get()};
   EXPECT_EQ(status, sourcemeta::hydra::http::Status::OK);
 }
 
 TEST(e2e_HTTP_Stream_1_1, no_callbacks_head) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::HEAD);
   const auto status{request.send().get()};
   EXPECT_EQ(status, sourcemeta::hydra::http::Status::OK);
 }
 
 TEST(e2e_HTTP_Stream_1_1, get_empty) {
-  sourcemeta::hydra::http::ClientStream request{std::string{HTTP_BASE_URL()} +
-                                                "/empty"};
+  sourcemeta::hydra::http::ClientStream request{
+      std::string{HTTPCLIENT_BASE_URL()} + "/empty"};
   request.method(sourcemeta::hydra::http::Method::GET);
 
   std::vector<sourcemeta::hydra::http::Status> statuses;
@@ -88,7 +88,7 @@ TEST(e2e_HTTP_Stream_1_1, get_empty) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, unsigned_long_header) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
   const unsigned long number{11};
   request.header("x-foo", number);
@@ -111,8 +111,8 @@ TEST(e2e_HTTP_Stream_1_1, unsigned_long_header) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, post_empty) {
-  sourcemeta::hydra::http::ClientStream request{std::string{HTTP_BASE_URL()} +
-                                                "/empty"};
+  sourcemeta::hydra::http::ClientStream request{
+      std::string{HTTPCLIENT_BASE_URL()} + "/empty"};
   request.method(sourcemeta::hydra::http::Method::POST);
 
   std::vector<sourcemeta::hydra::http::Status> statuses;
@@ -153,8 +153,8 @@ TEST(e2e_HTTP_Stream_1_1, post_empty) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, get_root_foo_bar) {
-  sourcemeta::hydra::http::ClientStream request{std::string{HTTP_BASE_URL()} +
-                                                "/foo/bar"};
+  sourcemeta::hydra::http::ClientStream request{
+      std::string{HTTPCLIENT_BASE_URL()} + "/foo/bar"};
   request.method(sourcemeta::hydra::http::Method::GET);
   request.header("X-Code", 400);
 
@@ -196,7 +196,7 @@ TEST(e2e_HTTP_Stream_1_1, get_root_foo_bar) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, on_body_exception) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::POST);
 
   request.on_body([](const std::size_t) -> std::vector<std::uint8_t> {
@@ -207,7 +207,7 @@ TEST(e2e_HTTP_Stream_1_1, on_body_exception) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, post_root_hello_world) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::POST);
 
   std::vector<sourcemeta::hydra::http::Status> statuses;
@@ -251,7 +251,7 @@ TEST(e2e_HTTP_Stream_1_1, post_root_hello_world) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, put_root_hello_world) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::PUT);
 
   std::vector<sourcemeta::hydra::http::Status> statuses;
@@ -295,7 +295,7 @@ TEST(e2e_HTTP_Stream_1_1, put_root_hello_world) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, get_root_hello_world) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::GET);
 
   std::vector<sourcemeta::hydra::http::Status> statuses;
@@ -339,7 +339,7 @@ TEST(e2e_HTTP_Stream_1_1, get_root_hello_world) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, post_root_empty_body) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::POST);
 
   std::vector<sourcemeta::hydra::http::Status> statuses;
@@ -376,7 +376,7 @@ TEST(e2e_HTTP_Stream_1_1, post_root_empty_body) {
 }
 
 TEST(e2e_HTTP_Stream_1_1, upload_chunked_encoding_by_default) {
-  sourcemeta::hydra::http::ClientStream request{HTTP_BASE_URL()};
+  sourcemeta::hydra::http::ClientStream request{HTTPCLIENT_BASE_URL()};
   request.method(sourcemeta::hydra::http::Method::PUT);
 
   std::istringstream request_body{"hello world"};
